feat: move About content into Home
